requirements.txt
# Список установленных модулей
# При клонировании репозитория на другой компьютер или сервер выполните
# (предварительно создав и активировав нужное виртуальное окружение):
# pip install -r requirements.txt
asgiref==3.3.1
colorame==0.4.4
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.1
loguru==0.5.3
Markdown==3.3.1
pip==20.2.3
psycopg2-binary==2.8.6
pytz==2020.4
setuptools==50.3.0
sqlparse==0.4.1
wheel==0.35.1
win32-setctime==1.0.3